move code to its own app (course_overview)
