+++ openedx/core/djangoapps/content/course_overview/tests.py
@@ -1,0 +1,1 @@
+# TODO me: write tests
