+++ openedx/core/djangoapps/content/course_structures/signals.py
@@ -1,7 +1,6 @@
 from django.dispatch.dispatcher import receiver
 
 from xmodule.modulestore.django import SignalHandler
-from .models import CourseOverviewFields
 
 @receiver(SignalHandler.course_published)
 def listen_for_course_publish(sender, course_key, **kwargs):  # pylint: disable=unused-argument
@@ -11,6 +10,3 @@ def listen_for_course_publish(sender, course_key, **kwargs):  # pylint: disable=
     # Note: The countdown=0 kwarg is set to to ensure the method below does not attempt to access the course
     # before the signal emitter has finished all operations. This is also necessary to ensure all tests pass.
     update_course_structure.apply_async([unicode(course_key)], countdown=0)
-
-    # TODO me: confirm that all cache invalidations will be hit by this signal
-    CourseOverviewFields.objects.filter(id=course_key).delete()
